requirements.txt
streamlit
python-dotenv
crewai
langchain-google-genai
